requirements.txt
streamlit
facenet-pytorch
torch
torchvision
numpy
pillow
cupy
mpi4py